python/requirements.txt
requests
python-socketio[client]
python-dotenv
